\name{DataStructure}
\alias{DataStructure}
\title{ Data structure of LGS output }
\description{
  \code{summaryLGS} provides summary statistics about the allele ratio and populations of an LGS object.
}
\author{ Dario Beraldi <\email{dario.beraldi@ed.ac.uk}> }
% Add one or more standard keywords, see file 'KEYWORDS' in the
% R documentation directory.
\keyword{ manip }

